requirements.txt
psycopg2-binary==2.7.5
Django==2.1.4
flake8==3.6.0
pylint==2.2.2
django-extensions==2.1.4
ipython==7.2.0
Pillow==5.3.0
ipdb
django-debug-toolbar==1.11
django-braces==1.13.0
django-embed-video==1.2.0
python-memcached==1.59
django-memcache-status==1.3
djangorestframework==3.9.1
uwsgi==2.0.17
# social-auth-app-django==3.1.0
# sorl-thumbnail==12.5.0
# mod_wsgi==4.6.5
# redis==3.0.1
# django-taggit==0.23.0
# Markdown==3.0.1
# django-haystack==2.8.1
# pysolr==3.8.1
